Improved error loggging

+++ dao_interface.go
@@ -8,10 +8,10 @@ type Entity interface {
 	UnmarshallHTTP(*http.Request) error
 
 	/* Populates the struct from DB with ID*/
-	GetFromDb(id int) error
+	GetFromDb(id int64) error
 
-	/* Adds the entity to the DB */
-	PutInDb() (int, error)
+	/* Adds the entity to the DB and return ID*/
+	PutInDb() (int64, error)
 
 	/* Updates the db with new values from struct */
 	UpdateInDb() error
